requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
apache-beam==2.64.0
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bokeh==3.7.3
cachetools==5.5.2
Cartopy==0.23.0
certifi==2025.4.26
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.3.2
crcmod==1.7
cycler==0.12.1
dask==2024.4.2
dc-stat-think==1.1.1
debugpy==1.8.14
decorator==5.2.1
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastavro==1.11.1
fasteners==0.19
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2024.3.1
gcsfs==2024.3.1
google-api-core==2.25.0rc1
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.65.5
h5netcdf==1.3.0
h5py==3.13.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_metadata==8.7.0
ipykernel==6.29.4
ipython==8.36.0
isodisreg @ git+https://github.com/evwalz/isodisreg.git@c293eb3a126f200bf4894ecbab6157596c2a395e
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpickle==3.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llvmlite==0.44.0
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
numba==0.61.2
numcodecs==0.13.1
numpy==2.2.6
oauthlib==3.2.2
objsize==0.7.1
opt_einsum==3.4.0
orjson==3.10.18
osqp==1.0.4
packaging==25.0
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
progressbar==2.5
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydot==1.4.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.13.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rechunker==0.5.4
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.0
scoringrules==0.8.0
seaborn==0.13.2
Shapely==1.8.5.post1
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uroc @ git+https://github.com/evwalz/urocc.git@13da5cf5c8d57af4a81cb1f78dd51a088f370d59
wcwidth==0.2.13
weatherbench2 @ git+https://github.com/google-research/weatherbench2.git@d29e2692ecce309b52407d7914af3baa62bbe2b9
xarray==2025.4.0
xarray-beam==0.8.0
xyzservices==2025.4.0
yarl==1.20.0
zarr==2.17.2
zipp==3.22.0
zstandard==0.23.0
